Refactor workflow trigger dialog to use flow controller hook
